requirements.txt
streamlit
tensorflow
opencv-python-headless
scikit-learn
numpy
pillow
librosa
soundfile
requests
pydub
openai-whisper @ git+https://github.com/openai/whisper.git
ffmpeg-python